backend/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
python-jose==3.3.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==2.0.2
